# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |-
    WASAText messaging platform API.
    Connect with your friends effortlessly using WASAText! Send and receive messages,
    whether one-on-one or in groups, all from the convenience of your PC.
  version: "1.0.0"
  contact:
    name: WASAText Team
    email: support@wasatext.com
...
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
paths:
  /session:
    post:
      tags: ["login"]
      summary: Logs in the user
      description: |-
        If the user does not exist, it will be created,
        and an identifier is returned.
        If the user exists, the user identifier is returned.
      operationId: doLogin
      security: []
      requestBody:
        description: User details
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Login request payload
              properties:
                name:
                  type: string
                  description: Username for login
                  example: Maria
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                type: object
                description: Login response containing user identifier
                properties:
                  identifier:
                    type: string
                    description: Unique user identifier for authentication
                    example: "abcdef012345"
                    minLength: 1
                    maxLength: 255
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/username:
    parameters:
      - name: userId
        in: path
        required: true
        description: Unique identifier of the user
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "abcdef012345"
    put:
      tags: ["user"]
      summary: Update user's username
      description: Updates the current user's username if not already taken
      operationId: setMyUserName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Username update request
              properties:
                name:
                  type: string
                  description: New username for the user
                  minLength: 3
                  maxLength: 16
                  pattern: '^[a-zA-Z0-9_-]+$'
                  example: "NewUsername"
              required:
                - name
      responses:
        '204':
          description: Username updated successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/photo:
    parameters:
      - name: userId
        in: path
        required: true
        description: Unique identifier of the user
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "abcdef012345"
    put:
      tags: ["user"]
      summary: Set user profile photo
      description: Uploads and sets the user's profile photo
      operationId: setMyPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Photo upload request
              properties:
                photo:
                  type: string
                  format: binary
                  description: Profile photo file
              required:
                - photo
      responses:
        '204':
          description: Photo updated successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}/conversations:
    parameters:
      - name: userId
        in: path
        required: true
        description: Unique identifier of the user
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "abcdef012345"
    get:
      tags: ["conversations"]
      summary: Get user's conversations
      description: Retrieves all conversations for the authenticated user
      operationId: getMyConversations
      responses:
        '200':
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                description: Array of conversation summaries
                minItems: 0
                maxItems: 10000
                items:
                  $ref: '#/components/schemas/ConversationSummary'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}:
    parameters:
      - name: conversationId
        in: path
        required: true
        description: Unique identifier of the conversation
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "conv123456"
    get:
      tags: ["conversations"]
      summary: Get conversation messages
      description: Retrieves all messages from a specific conversation
      operationId: getConversation
      responses:
        '200':
          description: Conversation messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Conversation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /conversations/{conversationId}/messages:
    parameters:
      - name: conversationId
        in: path
        required: true
        description: Unique identifier of the conversation
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "conv123456"
    post:
      tags: ["messages"]
      summary: Send a message
      description: Sends a new message to a conversation
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Message content for text messages
              properties:
                content:
                  type: string
                  description: Text content of the message
                  example: "Hello there!"
                  minLength: 1
                  maxLength: 5000
                messageType:
                  type: string
                  description: Type of message being sent
                  enum: ["text", "photo"]
                  example: "text"
                replyTo:
                  type: string
                  description: ID of message being replied to
                  example: "msg789"
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 1
                  maxLength: 255
              required:
                - content
                - messageType
          multipart/form-data:
            schema:
              type: object
              description: Message content for photo messages
              properties:
                photo:
                  type: string
                  format: binary
                  description: Photo file to send
                messageType:
                  type: string
                  description: Type of message being sent
                  enum: ["photo"]
                replyTo:
                  type: string
                  description: ID of message being replied to
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 1
                  maxLength: 255
              required:
                - photo
                - messageType
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/{messageId}/forward:
    parameters:
      - name: messageId
        in: path
        required: true
        description: Unique identifier of the message to forward
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "msg123456"
    post:
      tags: ["messages"]
      summary: Forward a message
      description: Forwards an existing message to another conversation
      operationId: forwardMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Forward message request
              properties:
                targetConversationId:
                  type: string
                  description: ID of the conversation to forward the message to
                  example: "conv789012"
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 1
                  maxLength: 255
              required:
                - targetConversationId
      responses:
        '201':
          description: Message forwarded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/{messageId}/comments:
    parameters:
      - name: messageId
        in: path
        required: true
        description: Unique identifier of the message
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "msg123456"
    post:
      tags: ["messages"]
      summary: React to a message
      description: Adds an emoji reaction to a message
      operationId: commentMessage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Reaction request
              properties:
                emoji:
                  type: string
                  description: Emoji to react with
                  example: "👍"
                  minLength: 1
                  maxLength: 10
              required:
                - emoji
      responses:
        '201':
          description: Reaction added successfully
          content:
            application/json:
              schema:
                type: object
                description: Reaction confirmation
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful
                    example: true
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: ["messages"]
      summary: Remove reaction from message
      description: Removes the user's reaction from a message
      operationId: uncommentMessage
      responses:
        '204':
          description: Reaction removed successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message or reaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /messages/{messageId}:
    parameters:
      - name: messageId
        in: path
        required: true
        description: Unique identifier of the message
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "msg123456"
    delete:
      tags: ["messages"]
      summary: Delete a message
      description: Deletes a message sent by the current user
      operationId: deleteMessage
      responses:
        '204':
          description: Message deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Cannot delete message from another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Message not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/members:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Unique identifier of the group
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "group123456"
    post:
      tags: ["groups"]
      summary: Add user to group
      description: Adds a user to an existing group
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Add user to group request
              properties:
                userId:
                  type: string
                  description: ID of the user to add to the group
                  example: "user789012"
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 1
                  maxLength: 255
              required:
                - userId
      responses:
        '201':
          description: User added to group successfully
          content:
            application/json:
              schema:
                type: object
                description: Add member confirmation
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful
                    example: true
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/members/{userId}:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Unique identifier of the group
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "group123456"
      - name: userId
        in: path
        required: true
        description: Unique identifier of the user to remove
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "user789012"
    delete:
      tags: ["groups"]
      summary: Leave group
      description: Removes the current user from a group
      operationId: leaveGroup
      responses:
        '204':
          description: Left group successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/name:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Unique identifier of the group
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "group123456"
    put:
      tags: ["groups"]
      summary: Set group name
      description: Updates the name of a group
      operationId: setGroupName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Group name update request
              properties:
                name:
                  type: string
                  description: New name for the group
                  minLength: 1
                  maxLength: 50
                  pattern: '^.*$'
                  example: "My Friends Group"
              required:
                - name
      responses:
        '204':
          description: Group name updated successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /groups/{groupId}/photo:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Unique identifier of the group
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "group123456"
    put:
      tags: ["groups"]
      summary: Set group photo
      description: Uploads and sets the group's photo
      operationId: setGroupPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              description: Group photo upload request
              properties:
                photo:
                  type: string
                  format: binary
                  description: Group photo file
              required:
                - photo
      responses:
        '204':
          description: Group photo updated successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Use the user identifier returned from login
  schemas:
    Error:
      type: object
      description: Error response object
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Error message"
          minLength: 1
          maxLength: 1000
        code:
          type: string
          description: Machine-readable error code
          example: "ERROR_CODE"
          pattern: '^[A-Z_]+$'
          minLength: 1
          maxLength: 50
      required:
        - message
    User:
      type: object
      description: User information object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "user123456"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        username:
          type: string
          description: User's display name
          example: "Maria"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 3
          maxLength: 16
        photoUrl:
          type: string
          description: URL to user's profile photo
          example: "https://api.wasatext.com/photos/user123456.jpg"
          minLength: 1
          maxLength: 2048
      required:
        - id
        - username
    ConversationSummary:
      type: object
      description: Summary of a conversation for list view
      properties:
        id:
          type: string
          description: Unique conversation identifier
          example: "conv123456"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        name:
          type: string
          description: Display name of the conversation
          pattern: '^[a-zA-Z]+$'
          example: "Maria"
          minLength: 1
          maxLength: 100
        photoUrl:
          type: string
          description: URL to conversation photo
          example: "https://api.wasatext.com/photos/conv123456.jpg"
          minLength: 1
          maxLength: 2048
        lastMessage:
          $ref: '#/components/schemas/MessagePreview'
        lastActivity:
          type: string
          format: date-time
          description: Timestamp of last activity in the conversation
          example: "2025-09-24T14:30:00Z"
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|([+-]\d{2}:\d{2}))$'
        isGroup:
          type: boolean
          description: Whether this is a group conversation
          example: false
        unreadCount:
          type: integer
          description: Number of unread messages
          example: 3
          minimum: 0
          maximum: 99999
      required:
        - id
        - name
        - lastActivity
        - isGroup
        - unreadCount
    MessagePreview:
      type: object
      description: Preview of the last message in a conversation
      properties:
        content:
          type: string
          description: Text content or description of the message
          example: "Hello there!"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 0
          maxLength: 5000
        messageType:
          type: string
          description: Type of the message
          enum: ["text", "photo"]
          example: "text"
        senderName:
          type: string
          description: Name of the message sender
          example: "Maria"
          pattern: '^[a-zA-Z]+$'
          minLength: 1
          maxLength: 100
      required:
        - messageType
    Conversation:
      type: object
      description: Full conversation with all messages
      properties:
        id:
          type: string
          description: Unique conversation identifier
          example: "conv123456"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        name:
          type: string
          description: Display name of the conversation
          pattern: '^[a-zA-Z]+$'
          example: "Maria"
          minLength: 1
          maxLength: 100
        photoUrl:
          type: string
          description: URL to conversation photo
          example: "https://api.wasatext.com/photos/conv123456.jpg"
          minLength: 1
          maxLength: 2048
        isGroup:
          type: boolean
          description: Whether this is a group conversation
          example: false
        members:
          type: array
          description: List of conversation members
          minItems: 0
          maxItems: 1000
          items:
            $ref: '#/components/schemas/User'
        messages:
          type: array
          description: List of messages in the conversation
          minItems: 0
          maxItems: 100000
          items:
            $ref: '#/components/schemas/Message'
      required:
        - id
        - name
        - isGroup
        - messages
    Message:
      type: object
      description: A single message in a conversation
      properties:
        id:
          type: string
          description: Unique message identifier
          example: "msg123456"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        content:
          type: string
          description: Text content of the message
          example: "Hello there!"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 0
          maxLength: 5000
        messageType:
          type: string
          description: Type of the message
          enum: ["text", "photo"]
          example: "text"
        timestamp:
          type: string
          format: date-time
          description: When the message was sent
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|([+-]\d{2}:\d{2}))$'
          example: "2025-09-24T14:30:00Z"
        senderId:
          type: string
          description: ID of the user who sent the message
          example: "user123456"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        senderName:
          type: string
          description: Display name of the sender
          example: "Maria"
          pattern: "^[a-zA-Z]+$"
          minLength: 1
          maxLength: 100
        delivered:
          type: boolean
          description: Whether the message has been delivered
          example: true
        read:
          type: boolean
          description: Whether the message has been read
          example: false
        replyTo:
          type: string
          description: ID of the message this is replying to
          example: "msg789012"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        reactions:
          type: array
          description: List of reactions to this message
          minItems: 0
          maxItems: 1000
          items:
            $ref: '#/components/schemas/Reaction'
        photoUrl:
          type: string
          description: URL to photo if messageType is photo
          example: "https://api.wasatext.com/photos/msg123456.jpg"
          minLength: 1
          maxLength: 2048
      required:
        - id
        - messageType
        - timestamp
        - senderId
        - delivered
        - read
    Reaction:
      type: object
      description: An emoji reaction to a message
      properties:
        userId:
          type: string
          description: ID of the user who reacted
          example: "user123456"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        userName:
          type: string
          description: Display name of the user who reacted
          example: "Maria"
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 100
        emoji:
          type: string
          description: The emoji used for the reaction
          example: "👍"
          pattern: '^.+$'
          minLength: 1
          maxLength: 10
        timestamp:
          type: string
          format: date-time
          description: When the reaction was added
          example: "2025-09-24T14:30:00Z"
          pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|([+-]\d{2}:\d{2}))$'
      required:
        - userId
        - userName
        - emoji
        - timestamp